requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
billiard==4.2.2
binaryornot==0.4.4
blinker==1.4
cachetools==5.5.2
celery==5.5.3
certifi==2020.6.20
channels==4.2.2
channels_redis==4.3.0
chardet==4.0.0
charset-normalizer==2.0.6

click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.4
constantly==23.10.4
cookiecutter==2.6.0
crispy-tailwind==1.0.3
cryptography==3.4.8
cupshelpers==1.0
daphne==4.2.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
dj-database-url==3.0.0
Django==5.2.2
django-browser-reload==1.18.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-htmx==1.23.0
django-tailwind==4.0.1
djangorestframework==3.16.1
exceptiongroup==1.3.0
ffmpeg-python==0.2.0
frozenlist==1.2.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gyp==0.1
h11==0.16.0
hidpidaemon==18.4.6
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==24.7.2
jeepney==0.7.1
Jinja2==3.1.6
kernelstub==3.1.4
keyring==23.5.0
kombu==5.5.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.1.1
multidict==5.1.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==21.3
pillow==11.3.0
pop-transition==1.1.2
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.42.5
proton-vpn-gtk-app==4.9.7
proton-vpn-network-manager==0.12.14
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.7
pydantic_core==2.33.2
pydbus==0.6.0
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-gnupg==0.4.8
python-slugify==8.0.4
python-telegram-bot==22.3
aiogram>=3.0.0
gunicorn>=21.0.0
python-whois==0.9.5
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==6.4.0
replicate==1.0.7
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
rich==14.0.0
rsa==4.9.1
SecretStorage==3.3.1
sentry-sdk==1.4.3
service-identity==24.2.0
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
systemd-python==234
text-unidecode==1.3
tomli==2.2.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
uritemplate==4.2.0
urllib3==1.26.5
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.14
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
zope.interface==7.2
